support for qwen3, qwen2 added, torch(bf16) eval py, upload.py added

+++ sbvr/kernels/sbvr_prefill_kernel.cu
@@ -180,9 +180,9 @@ __device__ void store_frag_as_fp16_guarded(
 
  template <typename RIndexT, int RNumSums>
  __global__ void sbvr_prefill_mm_T(
-    __nv_bfloat16* __restrict__ x,
-    uint32_t* r_bvr, RIndexT* r_coeff_idx, __nv_bfloat16* __restrict__ r_coeff_cache,
-    __nv_bfloat16* __restrict__ bias, __nv_bfloat16* __restrict__ out,
+    half* __restrict__ x,
+    uint32_t* r_bvr, RIndexT* r_coeff_idx, half* __restrict__ r_coeff_cache,
+    half* __restrict__ bias, half* __restrict__ out,
     int M, int N, int K)
 {
     // 1. CTA indices and other important indices
@@ -266,11 +266,11 @@ __device__ void store_frag_as_fp16_guarded(
  *  Note that in prefill kernel, K is not divided by 32 originally
  * ========================================================================= */
 void launch_prefill_sbvr_kernel(
-    __nv_bfloat16* x,
-    uint32_t* r_bvr, void* r_coeff_idx, __nv_bfloat16* r_coeff_cache,
+    half* x,
+    uint32_t* r_bvr, void* r_coeff_idx, half* r_coeff_cache,
     int r_num_sums,
     int r_cache_size,
-    __nv_bfloat16* bias, __nv_bfloat16* out,
+    half* bias, half* out,
     int M, int N, int K)
 {
     // 1. determine the type of r_coeff_idx
